+++ exsu/report.py
@@ -99,13 +99,13 @@ class Report:
 
     def set_output_dir(self, outputdir):
         """
-        Set the output dir and create the directory if it does not exist yet.
+        Set the output dir with expand user.
         :param outputdir:
         :return:
         """
         self.outputdir = Path(outputdir).expanduser()
-        if not op.exists(self.outputdir):
-            os.makedirs(self.outputdir)
+        # if not op.exists(self.outputdir):
+        #     os.makedirs(self.outputdir)
 
     def set_show(self, show):
         self.show = show
@@ -152,7 +152,7 @@ class Report:
 
     def dump(self):
         if self.outputdir is not None:
-            pth = self._imjoin(self.outputdir, self.spreadsheet_fn)
+            pth = self.join_output_dir(self.outputdir, self.spreadsheet_fn)
             logger.debug(f"Saving to file {pth}")
             ppth = Path(pth)
             if not ppth.parent.exists():
@@ -204,7 +204,7 @@ class Report:
             if self._is_under_level(level):
                 import matplotlib.pyplot as plt
 
-                fn = self._imjoin(self.outputdir, filename + ext)
+                fn = self.join_output_dir(self.outputdir, filename + ext)
                 logger.debug(f"write to file: {fn}")
                 plt.imsave(fn, arr, **kwargs)
 
@@ -212,15 +212,17 @@ class Report:
                 with warnings.catch_warnings():
                     warnings.filterwarnings("ignore", ".*low contrast image.*")
                     # warnings.simplefilter("low contrast image")
-                    fn = self._imjoin(self.outputdir, filename + "_skimage" + ext)
+                    fn = self.join_output_dir(self.outputdir, filename + "_skimage" + ext)
                     logger.debug(f"write to file: {fn}")
                     skimage.io.imsave(fn, k * arr, *kwargs_skimage)
             if self._is_under_level(level_npz):
                 self._save_arr(base_fn, arr)
 
-    def _imjoin(self, *args, **kwargs):
+    def join_output_dir(self, *args, **kwargs):
         if self.outputdir is None:
-            raise ValueError("Outputdir have to be set before image is saved.")
+            raise ValueError("Outputdir have to be set before file is saved.")
+        if not op.exists(self.outputdir):
+            os.makedirs(self.outputdir)
         return op.join(*args, **kwargs)
 
     def _save_arr(self, base_fn, arr: np.ndarray):
@@ -231,7 +233,7 @@ class Report:
         :return:
         """
         filename, ext = op.splitext(base_fn)
-        fn = self._imjoin(self.outputdir, filename + ".npz")
+        fn = self.join_output_dir(self.outputdir, filename + ".npz")
         np.savez_compressed(fn, arr=arr)
         self.imgs[base_fn] = fn
 
@@ -262,7 +264,7 @@ class Report:
             plt.imshow(arr)
             plt.colorbar()
             if self.save:
-                fn = self._imjoin(self.outputdir, filename + "" + ext)
+                fn = self.join_output_dir(self.outputdir, filename + "" + ext)
                 plt.savefig(fn)
             if self.show:
                 plt.show()
@@ -291,7 +293,7 @@ class Report:
             if ext == "":
                 ext = ".png"
             if self.save:
-                fn = self._imjoin(self.outputdir, filename + "" + ext)
+                fn = self.join_output_dir(self.outputdir, filename + "" + ext)
                 plt.savefig(fn)
                 # self.imgs[base_fn] = [fn]