+++ changelog.md
@@ -1,5 +1,8 @@
 # Changelog
 
+Actual version 0.4.1
+
+## 0.5
 
  * [Changed] Additional spreadsheet saved without index to be compatible with common spreadsheet
  * [Added] Find if of closest points in the dataframe